# Optimize fits.Header parsing