@MSeifert04 2019-02-16
Could you share the fits or the code to produce one that's like it?