@saimn 2019-02-17
I cannot share this specific file, I could build a similar one but this should noticeable for any file with many keywords (and not specifically for files with many HDUs - I updated the issue title -, this will be a following PR). So it would be useful if you can test on other files ;)

Here is the profiler result, a lot a time spent in `Header.append` (hence the new `_fromcards`) and in keyword parsing:
![capture d ecran de 2019-02-17 21-49-28](https://user-images.githubusercontent.com/311639/52919020-152f1580-32fe-11e9-8c4e-f52393f4d245.png)