@saimn 2019-02-18
Ok, I reverted the change for the `clear` call, I still find it weird to have to call a `clear` method as the first line of `__init__` but this is not a big deal. I have another branch with a dozen commits based on this one so I would prefer not having to split things now, and the cosmetic changes are rather limited.