@pllim 2019-02-19
By the way, do we need a benchmark for this over at https://github.com/astropy/astropy-benchmarks ?