@saimn 2019-02-19
Thanks for the reviews.

@mhvk - I don't like it too, but the overhead of a function call is not negligible when it is done for thousands of keywords. Would be great if Python could optimize this itself.

@pllim - Yes, it would be great. I will see of I can add some benchmarks.